Change to use `Duration` for block intervals
This way, the user may specify the interval in any prefix they like.
Since `std::thread::sleep` takes a `Duration` instance, we previously
converted the seconds to one anyway - why not remove this step? Less
code, more flexibility, more good. :)

This is the implementation of a suggestion by @gevhaz on GitHub. Thanks!